redorganized Resourcemanager

+++ src/items/Weapon.java
@@ -1,14 +1,14 @@
 package items;
 
 public class Weapon extends Item {
-	private double damage;
+	private int damage;
 
-	public Weapon(String uniqueName, String name, int value, double damage, int spwanChance) {
+	public Weapon(String uniqueName, String name, int value, int damage, int spwanChance) {
 		super(uniqueName, name, value, spwanChance);
 		this.damage = damage;
 	}
 
-	public double getDamage() {
+	public int getDamage() {
 		return damage;
 	}
